requirements.txt
pandas>=2.0.0
requests>=2.25.0
tqdm>=4.65.0
openpyxl>=3.0.0
pyyaml>=6.0.2